extract span overview list into separate component
